pyproject.toml
[project]
name = "webvoyager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "openai>=2.6.1",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "selenium>=4.38.0",
]
